// src/app/features/products/interfaces/IGetAllProducts.ts
export interface Metadata {
	currentPage: number;
	numberOfPages: number;
	limit: number;
	nextPage: number;
}

export interface Subcategory {
	_id: string;
	name: string;
	slug: string;
	category: string;
}

export interface Category {
	_id: string;
	name: string;
	slug: string;
	image: string;
}

export interface Brand {
	_id: string;
	name: string;
	slug: string;
	image: string;
}

export interface IProductsData {
	sold: number;
	images: string[];
	subcategory: Subcategory[];
	ratingsQuantity: number;
	_id: string;
	title: string;
	slug: string;
	description: string;
	quantity: number;
	price: number;
	imageCover: string;
	category: Category;
	brand: Brand;
	ratingsAverage: number;
	createdAt: string;
	updatedAt: string;
	__v: number;
	reviews: any[];
	id: string;
}

export interface IGetAllProducts {
	results: number;
	metadata: Metadata;
	data: IProductsData[];
}